A function that prints all the elements of a dlistint_t list

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -1,7 +1,8 @@
 #include "lists.h"
 #include <stdio.h>
+#include <stddef.h>
 
-/*
+/**
  * print_dlistint - This will print all the elements of the
  * function list
  *
@@ -11,7 +12,7 @@
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t counter = 0;
+	int counter = 0;
 
 	while (h != NULL)
 	{
